+++ examples/start-up-timeout.R
@@ -1,0 +1,25 @@
+# with timeout ------------------------------------------------------------
+
+library(shiny)
+library(shinybusy)
+
+ui <- fluidPage(
+
+  busy_start_up(
+    loader = spin_epic("orbit", color = "#FFF"),
+    text = "Loading...",
+    timeout = 1500,
+    color = "#FFF",
+    background = "#112446"
+  ),
+
+  tags$h1("Ready to play!", class = "text-center")
+
+)
+
+server <- function(input, output, session) {
+
+}
+
+if (interactive())
+  shinyApp(ui, server)